Tests changed to test daos

+++ src/main/java/com/epam/note/persistence/NoteRepository.java
@@ -7,6 +7,6 @@ public interface NoteRepository extends JpaRepository<Note, Long> {
 
 
     Note save(Note note);
-    Note getById(int i);
+    Note getById(long i);
 
 }
